requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
python-docx==0.8.11
python-pptx==0.6.23
requests==2.31.0
bcrypt==4.0.1